+++ include/openrave/kinbody.h
@@ -47,6 +47,7 @@ enum GeometryType {
     GT_Cage=6, ///< a container shaped geometry with removable side walls. The side walls can be on any of the four sides. The origin is at the bottom of the base. The inner volume of the cage is measured from the base to the highest wall.
     GT_CalibrationBoard=7, ///< a box shaped geometry with grid of cylindrical dots of two sizes. The dots are always on the +z side of the box and are oriented towards z-axis.
     GT_Axial = 8, ///< a geometry defined by many slices along an axis, oriented towards z-axis
+    GT_ConicalFrustum = 9, ///< a geometry defined by a conical frustum, oriented towards z-axis
 };
 
 enum DynamicsConstraintsType {
@@ -394,18 +395,19 @@ public:
         inline dReal GetSphereRadius() const {
             return _vGeomData.x;
         }
-        /// \deprecated use GetCylinderTopRadius and GetCylinderBottomRadius
-        dReal GetCylinderRadius() const;
-        inline dReal GetCylinderTopRadius() const {
+        inline dReal GetCylinderRadius() const {
             return _vGeomData.x;
         }
         inline dReal GetCylinderHeight() const {
             return _vGeomData.y;
         }
-        inline dReal GetCylinderBottomRadius() const {
-            if (_vGeomData.z == 0) {
-                return _vGeomData.x;
-            }
+        inline dReal GetConicalFrustumTopRadius() const {
+            return _vGeomData.x;
+        }
+        inline dReal GetConicalFrustumBottomRadius() const {
+            return _vGeomData.y;
+        }
+        inline dReal GetConicalFrustumHeight() const {
             return _vGeomData.z;
         }
         inline const Vector& GetBoxExtents() const {
@@ -451,8 +453,9 @@ public:
         }
 
         ///< for sphere it is radius
-        ///< for cylinder, first 3 values are topRadius, height and bottomRadius
+        ///< for cylinder, first 2 values are radius and height
         ///< for trimesh, none
+        ///< for conical frustum, first 3 values are top radius, bottom radius, height
         /// for boxes, first 3 values are half extents. For containers, the first 3 values are the full outer extents.
         /// For GT_Cage, this is the base box extents with the origin being at the -Z center.
         Vector _vGeomData;
@@ -701,21 +704,24 @@ public:
         inline bool IsModifiable() const {
             return _info._bModifiable;
         }
+
         inline dReal GetSphereRadius() const {
-            return _info._vGeomData.x;
+            return _info.GetSphereRadius();
         }
-        /// \deprecated use GetCylinderTopRadius and GetCylinderBottomRadius
         inline dReal GetCylinderRadius() const {
             return _info.GetCylinderRadius();
-        }
-        inline dReal GetCylinderTopRadius() const {
-            return _info.GetCylinderTopRadius();
         }
         inline dReal GetCylinderHeight() const {
             return _info.GetCylinderHeight();
         }
-        inline dReal GetCylinderBottomRadius() const {
-            return _info.GetCylinderBottomRadius();
+        inline dReal GetConicalFrustumTopRadius() const {
+            return _info.GetConicalFrustumTopRadius();
+        }
+        inline dReal GetConicalFrustumBottomRadius() const {
+            return _info.GetConicalFrustumBottomRadius();
+        }
+        inline dReal GetConicalFrustumHeight() const {
+            return _info.GetConicalFrustumHeight();
         }
         inline const Vector& GetBoxExtents() const {
             return _info.GetBoxExtents();